requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
psycopg[binary]
requests
